requirements.txt
appdirs==1.4.3
asn1crypto==0.24.0
attrs==18.2.0
Babel==2.6.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cliff==2.14.0
cmd2==0.9.6
colorama==0.4.0
cryptography==2.4.1
debtcollector==1.20.0
decorator==4.3.0
dogpile.cache==0.6.7
filelock==3.0.10
idna==2.7
iso8601==0.1.12
jmespath==0.9.3
jsonpatch==1.23
jsonpointer==2.0
jsonschema==2.6.0
keystoneauth1==3.11.1
monotonic==1.5
msgpack==0.5.6
munch==2.3.2
netaddr==0.7.19
netifaces==0.10.7
openstacksdk==0.19.0
os-client-config==1.31.2
os-service-types==1.3.0
osc-lib==1.11.1
oslo.config==6.7.0
oslo.context==2.21.0
oslo.i18n==3.22.1
oslo.log==3.40.1
oslo.serialization==2.28.1
oslo.utils==3.37.1
pbr==5.1.1
pluggy==0.8.0
prettytable==0.7.2
py==1.7.0
pycparser==2.19
Pygments==2.2.0
pyinotify==0.9.6
pyOpenSSL==18.0.0
pyparsing==2.3.0
pyperclip==1.7.0
python-ceilometerclient==2.9.0
python-cinderclient==4.1.0
python-dateutil==2.7.5
python-designateclient==2.11.0
python-glanceclient==2.14.0
python-keystoneclient==3.18.0
python-neutronclient==6.11.0
python-novaclient==11.1.0
python-openstackclient==3.17.0
python-swiftclient==3.6.0
pytz==2018.7
PyYAML==3.13
requests==2.20.1
requestsexceptions==1.4.0
rfc3986==1.1.0
serpy==0.3.1
simplejson==3.16.0
six==1.11.0
stevedore==1.30.0
tabulate==0.8.2
termcolor==1.1.0
terminal==0.4.0
toml==0.10.0
tox==3.5.3
urllib3==1.24.1
virtualenv==16.1.0
warlock==1.3.0
wcwidth==0.1.7
wrapt==1.10.11
